Adding tests in testing suite with test that (#29)
Have added tests for the basic SIR running functions (deterministic and
stochastic) as well as plotting functions.

Future work will troubleshoot any remaining issues in tests and compare
to validation scripts. Both of these are repetitive and not DRY in some
ways and a discussion is needed to understand path forward on testing
vs. validation. Once this is decided, lessons learned in these testing
suites will be applied to test functions for the generalized modeling
functions and possibly helper functions.