Development of wrap_adaptivetau() + example-adaptivetau script (#94)
This PR addresses:


Issue #95 : develops a wrapper for [adaptivetau::ssa.exact()] and
[adaptivetau::ssa.adaptivetau()] and updates examples-adaptivetau.R to
use this

Notes to reviewer:
1) During development it became clear that a difference between this
wrapper and the wrap_gillespie() function was needed to handle the
initial values (init_vals) argument of the wrapper. In wrap_gillespie()
one could run x0 outside of the wrapper and use it as the init_vals
argument to wrap_gillespie() without error. However, in
wrap_adaptivetau() used here one needs the initial values that would
have been fed into define_initialstate instead. In the examples, x0, the
output of define_initialstate, has the modified states and when run this
causes the error. Perhaps this means a revisit or alteration to the
examples-gillespie.R and warp_gillespie() may be needed.

2) For the last model in the example script, the use of ssa.exact() took
notably longer to run 10 simulations than ssa.adaptivetau. Not sure what
is worth investigating there. A single simulation using ssa.exact() was
about 3-4 minutes. (Run times all in VSCode on GFE).