Adding testing.yaml to run pkg check during GHA (Issue #66) (#138)
Closes Issue #66 
Adds a package check to GitHub Actions. Although we routinely check the
package locally prior to PR, this protects us further against the case
where a user forgets to do so.

Happily this PR passes the new GHA check just fine and for the time
being only adds another minute (estimated) to the checks.

Based on the following:
https://github.com/cdcent/cfa-msm-mpox/blob/main/.github/workflows/testing.yaml