Set seed for stochastic simulation in test-sir-extinctionprobability-j-infections.R (#145)
Now that we have a testing.yaml that checks R packages and runs the test
suites as part of GHA, some stochastic tests have failed more frequently
due to values outside the tolerance of the test being stochastically
selected. The workaround has been to re-run the failed jobs on GH.

To reduce this happening, the options were to adjust the tolerance for
the test (eg make it larger) or to set a seed of a simulation that
worked. This PR does the latter.

This PR adds one line to set the seed on line 13 ahead of stochastic
simulation using wrap_adaptivetau in
test-sir-extinctionprobability-j-infections.R.