Functions to connect to griddleR for deterministic and stochastic models (Issue #135) (#141)
Fixes Issue #135 
First pass script to demonstrate linkages from a compiled compModels SIR
model to griddleR functionality for deterministic and stochastic models.
These include wrapper functions called double_wrap_ode() and
double_wrap_adaptivetau().

The griddleR::run() function needed minor modifications to be able to
pass additional arguments and set random seeds in stochastic models.
This is being addressed via the following PR in the griddleR repo:
https://github.com/CDCgov/griddleR/pull/11
Once landed, then I expect the double_wrap_ prefixed functions could
become standalone in the compModels/R and be used when called the
revised griddleR::run() function for users.

Note to reviewer: due to issues with local/remote tracking on this
branch (I lost the local branch somehow and when I brought it back down
main was ahead), I rebased locally to an updated main so 7 commits came
along that are not from this branch. The final two (looks like top two
below) are from this branch and intended to be reviewed as part of the
PR. The rest were already landed on main.

the draft:
[8f176d5](https://github.com/cdcent/cfa-gen-comp-model-tools/pull/141/commits/8f176d55606a9dbd2261a7aceab27eaab4d575e1)
the final:
[8d1e3ce](https://github.com/cdcent/cfa-gen-comp-model-tools/pull/141/commits/8d1e3cea6a15fc2bd1d540b2722fbbde436cf441)