2024 Measles Model with case ascertainment delay as a function fn, revised measles model script (#161)
This PR adds:
- the 2024 Measles Model with case ascertainment delay as a function
- a revised measles model script full-measles-chicago.2024.R
- 
Note to reviewer, the case ascertainment delay object obs_df has to be
plotted with a user-defined plot as plot_stoch_model expects a list of
dataframes and has code to handle plotting multiple simulations. So
plot_stoch_model() is only plotting the dynamics object (alldyn)
currently.

Request for reviewer: carefully review the timesteps and timeline
throughout.