Quick Fix #2 for PR #171 to address site build errors (#175)
This PR adds:
- Typo fix for empty2states.R Roxygen tag @family internal - "internal"
had a typo
- Re-ran devtools::document() to make sure man/empty2states.Rd reflected
typo fix
- ._pkgdown.yaml has_concept("internal", internal = TRUE) - added
internal = TRUE to have reference index show internal functions (eg
functions without @export in Roxygen). The reference_index does NOT
include these by default. See this
[link](https://pkgdown.r-lib.org/reference/build_reference.html#:~:text=By%20default%2C%20pkgdown%20will%20generate%20an%20index%20that,the%20.Rd%20is%20more%20recent%20than%20the%20.html.).
- Now the build works, and it will deploy once it hits main.

Lastly, there is an additional of the data/ directory to the .gitignore
file for this repository for easier git workflow.

Additional information for posterity:
I additionally tested this with @gvegayon on a different system as
initially these first three steps (which are the actual fix) did not
work locally on my machine (or several other attempts - you'll see in
commit history). We explored some farther out ideas such as the @keyword
internal Roxygen tag and other elements but both of us knew this wasn't
needed. We also explored if workflows/pkgdown.yaml required a specific
branch for correct deployment. We are both able to make the first step
in workflow work on this branch in GH now.