Benchmarking stochastic simulation package fns and compModels stochastic wrappers (Issue #99) (#169)
Resolves #99 
- Refreshed previous benchmarking work on stochastic simulation package
fns and compModels stochastic wrappers now that wrapper functions were
modified for improved speed (PR #139).
- Additionally, fixed typos in complex models and benchmarking, adjusted
the microbenchmark times argument to 10 (vs default 100) for faster
results viewing.
- Have not included any further runs on GillespieSSA::ssa() or
wrap_gillespie() but could add if we think this is needed.


*BELOW IS NOW RESOLVED, SEE COMMENTS*
**Reviewer request for help**:
- running more complex models than SIR with adaptivetau::ssa.adaptivetau
yields one of two errors (below; I have commented this code out). This
is not present when run in wrap_adaptivetau(). Can you help decipher
what is going on? There is either an issue in transition matrix or rate
function, and I think it's the former but how to resolve is not
immediately obvious.

I see this error 6 times:
```
Error: invalid transition specification
```
and this error 2 times:
```
Error in adaptivetau::ssa.adaptivetau(init.values = seir_init_vals, transitions = # as.matrix(seir_peter), : 
negative variable: 3 is -1 (check rate function and/or transition matrix)
```